omnia fixes
